+++ Codex/src/main/java/org/operatorfoundation/codex/symbols/CallNumber.kt
@@ -13,7 +13,7 @@ class CallNumber(val value: Char) : Symbol {
         )
         private val valueToChar = charToValue.map { (k, v) -> v to k }.toMap()
 
-        override fun size(): Int = 10
+        override fun size(): BigInteger = 10.toBigInteger()
 
         override fun encode(numericValue: BigInteger): CallNumber {
             val int = numericValue.toInt()
