+++ Codex/src/main/java/org/operatorfoundation/codex/symbols/Trinary.kt
@@ -7,7 +7,7 @@ class Trinary(val value: Int) : Symbol {
     companion object : SymbolFactory<Trinary> {
         const val MAX = 2
 
-        override fun size(): Int = MAX + 1
+        override fun size(): BigInteger = (MAX + 1).toBigInteger()
 
         /**
          * Encodes a numeric value to its byte array representation.
